+++ image/lib/src/pubspec.dart
@@ -16,17 +16,17 @@ class Pubspec {
   static const String issue_tracker =
       'https://github.com/DavBfr/dart_barcode/issues';
 
-  static const String versionFull = '1.0.3';
+  static const String versionFull = '1.1.0';
 
-  static const String version = '1.0.3';
+  static const String version = '1.1.0';
 
-  static const String versionSmall = '1.0';
+  static const String versionSmall = '1.1';
 
   static const int versionMajor = 1;
 
-  static const int versionMinor = 0;
+  static const int versionMinor = 1;
 
-  static const int versionPatch = 3;
+  static const int versionPatch = 0;
 
   static const int versionBuild = 0;
 
@@ -39,8 +39,8 @@ class Pubspec {
   };
 
   static const Map<dynamic, dynamic> dependencies = <dynamic, dynamic>{
-    'barcode': '^1.11.1',
-    'meta': '^1.1.5',
+    'barcode': '^1.14.0',
+    'meta': '^1.1.8',
     'image': '^2.1.0',
     'args': '^1.5.0',
   };